lazy all commands
