+++ packages/expo-cli/src/commands/loginAsync.ts
@@ -1,0 +1,3 @@
+import { login } from '../accounts';
+
+export const actionAsync = login;
