+++ packages/expo-cli/src/commands/prepare-detached-build.ts
@@ -1,21 +1,15 @@
 import type { Command } from 'commander';
-import { Detach } from 'xdl';
 
-type Options = {
-  platform?: string;
-  skipXcodeConfig: boolean;
-};
-
-async function action(projectRoot: string, options: Options) {
-  await Detach.prepareDetachedBuildAsync(projectRoot, options);
-}
+import { applyAsyncActionProjectDir } from './utils/applyAsyncAction';
 
 export default function (program: Command) {
-  program
-    .command('prepare-detached-build [path]')
-    .description('Prepare a detached project for building')
-    .helpGroup('internal')
-    .option('--platform [platform]', 'detached project platform')
-    .option('--skipXcodeConfig [bool]', '[iOS only] if true, do not configure Xcode project')
-    .asyncActionProjectDir(action);
+  applyAsyncActionProjectDir(
+    program
+      .command('prepare-detached-build [path]')
+      .description('Prepare a detached project for building')
+      .helpGroup('internal')
+      .option('--platform [platform]', 'detached project platform')
+      .option('--skipXcodeConfig [bool]', '[iOS only] if true, do not configure Xcode project'),
+    () => import('./prepareDetachedBuildAsync')
+  );
 }
