+++ packages/expo-cli/src/commands/prepareDetachedBuildAsync.ts
@@ -1,0 +1,10 @@
+import { Detach } from 'xdl';
+
+type Options = {
+  platform?: string;
+  skipXcodeConfig: boolean;
+};
+
+export async function actionAsync(projectRoot: string, options: Options) {
+  await Detach.prepareDetachedBuildAsync(projectRoot, options);
+}
