+++ packages/expo-cli/src/commands/url/urlApkAsync.ts
@@ -1,0 +1,3 @@
+import { logArtifactUrl } from './logArtifactUrl';
+
+export const actionAsync = logArtifactUrl('android');
