+++ packages/expo-cli/src/commands/utils/applyAsyncAction.ts
@@ -5,7 +5,9 @@ import { profileMethod } from '../utils/profileMethod';
 
 export function applyAsyncActionProjectDir<Options = Record<string, any>>(
   command: Command,
-  resolve: () => Promise<{ actionAsync: (projectRoot: string, options: Options) => Promise<void> }>,
+  resolve: () => Promise<{
+    actionAsync: (projectRoot: string, options: Options) => Promise<unknown>;
+  }>,
   settings?: { checkConfig?: boolean; skipSDKVersionRequirement?: true }
 ) {
   command.asyncActionProjectDir(async (projectRoot: string, options: Options) => {
@@ -14,11 +16,11 @@ export function applyAsyncActionProjectDir<Options = Record<string, any>>(
   }, settings);
 }
 
-export function applyAsyncAction<Options = Record<string, any>>(
+export function applyAsyncAction<Args = string[], Options = Record<string, any>>(
   command: Command,
-  resolve: () => Promise<{ actionAsync: (args: string[], options: Options) => Promise<void> }>
+  resolve: () => Promise<{ actionAsync: (args: Args, options: Options) => Promise<unknown> }>
 ) {
-  command.asyncAction(async (args: string[], options: Options) => {
+  command.asyncAction(async (args: Args, options: Options) => {
     const mod = await resolve();
     return profileMethod(mod.actionAsync)(args, options);
   });
@@ -26,7 +28,7 @@ export function applyAsyncAction<Options = Record<string, any>>(
 
 export function applyAnyAsyncAction<Options = Record<string, any>>(
   command: Command,
-  resolve: () => Promise<{ actionAsync: (options: Options) => Promise<void> }>
+  resolve: () => Promise<{ actionAsync: (options: Options) => Promise<unknown> }>
 ) {
   command.asyncAction(async (options: Options) => {
     assert(typeof options !== 'string', 'Unexpected string passed to command');
